requirements.txt
colorama==0.3.6
nose==1.3.7
py==1.4.31
pytest==2.8.6
python-dateutil==2.4.2
python-termstyle==0.1.10
pytz==2015.7
requests==2.9.1
rollbar==0.13.17
six==1.10.0
sniffer==0.3.5
